scripts/requirements.txt
diffusers>=0.21.0
transformers>=4.31.0
accelerate>=0.21.0
safetensors>=0.3.1
pillow
torch
torchvision
